+++ doc/release_notes.rst
@@ -4,7 +4,7 @@ Release Notes
 Upcoming Release
 ----------------
 
-* Numpy Array functions were disabled in the `Variable`, `Constraint` and `LinearExpression` class in order to ensure persistence as the class when multiplying with `pandas` and `numpy` objects.
+* Numpy ``__array_ufunc__`` was disabled in the `Variable`, `Constraint` and `LinearExpression` class in order to ensure persistence as the class when multiplying with `numpy` objects. As for pandas objects the issue https://github.com/pandas-dev/pandas/issues/45803 must be solved.
 * The `Variable` class got a new accessor `sol` which points to the optimal values if the underlying model was optimized.
 * The `Constraint` class got a new accessor `dual` which points to the dual values if tune underlying model was optimized and dual values are existent.